# perf: Reduce `yarn add/yarn remove` time by validating cache file checksums only once