--- a/packages/plugin-npm/sources/NpmSemverFetcher.ts
+++ b/packages/plugin-npm/sources/NpmSemverFetcher.ts
@@ -34,6 +34,7 @@ export class NpmSemverFetcher implements Fetcher {
       onHit: () => opts.report.reportCacheHit(locator),
       onMiss: () => opts.report.reportCacheMiss(locator, `${structUtils.prettyLocator(opts.project.configuration, locator)} can't be found in the cache and will be fetched from the remote registry`),
       loader: () => this.fetchFromNetwork(locator, opts),
+      skipIntegrityCheck: opts.skipIntegrityCheck,
     });
 
     return {