--- a/packages/plugin-patch/sources/PatchFetcher.ts
+++ b/packages/plugin-patch/sources/PatchFetcher.ts
@@ -25,6 +25,7 @@ export class PatchFetcher implements Fetcher {
       onHit: () => opts.report.reportCacheHit(locator),
       onMiss: () => opts.report.reportCacheMiss(locator, `${structUtils.prettyLocator(opts.project.configuration, locator)} can't be found in the cache and will be fetched from the disk`),
       loader: () => this.patchPackage(locator, opts),
+      skipIntegrityCheck: opts.skipIntegrityCheck,
     });
 
     return {