--- a/packages/yarnpkg-core/sources/Cache.ts
+++ b/packages/yarnpkg-core/sources/Cache.ts
@@ -134,16 +134,16 @@ export class Cache {
     }
   }
 
-  async fetchPackageFromCache(locator: Locator, expectedChecksum: string | null, {onHit, onMiss, loader}: {onHit?: () => void, onMiss?: () => void, loader?: () => Promise<ZipFS>}): Promise<[FakeFS<PortablePath>, () => void, string]> {
+  async fetchPackageFromCache(locator: Locator, expectedChecksum: string | null, {onHit, onMiss, loader, skipIntegrityCheck}: {onHit?: () => void, onMiss?: () => void, loader?: () => Promise<ZipFS>, skipIntegrityCheck: boolean}): Promise<[FakeFS<PortablePath>, () => void, string]> {
     const mirrorPath = this.getLocatorMirrorPath(locator);
 
     const baseFs = new NodeFS();
 
     const validateFile = async (path: PortablePath, refetchPath: PortablePath | null = null) => {
-      const actualChecksum = `${this.cacheKey}/${await cachedChecksumFile(path)}`;
+      const actualChecksum = (!skipIntegrityCheck || !expectedChecksum) ? `${this.cacheKey}/${await hashUtils.checksumFile(path)}` : expectedChecksum;
 
       if (refetchPath !== null) {
-        const previousChecksum = `${this.cacheKey}/${await cachedChecksumFile(refetchPath)}`;
+        const previousChecksum = (!skipIntegrityCheck || !expectedChecksum) ? `${this.cacheKey}/${await hashUtils.checksumFile(refetchPath)}` : expectedChecksum;
         if (actualChecksum !== previousChecksum) {
           throw new ReportError(MessageName.CACHE_CHECKSUM_MISMATCH, `The remote archive doesn't match the local checksum - has the local cache been corrupted?`);
         }
@@ -325,15 +325,4 @@ function getCacheKeyComponent(checksum: string) {
 function getHashComponent(checksum: string) {
   const split = checksum.indexOf(`/`);
   return split !== -1 ? checksum.slice(split + 1) : checksum;
-}
-
-const checksumCache = new Map<PortablePath, string>();
-
-async function cachedChecksumFile(path: PortablePath) {
-  let checksum = checksumCache.get(path);
-  if (!checksum) {
-    checksum = await hashUtils.checksumFile(path);
-    checksumCache.set(path, checksum);
-  }
-  return checksum;
 }