**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

Yarn computes cache file checksums during `Fetch step` and then computes the same cache file checksums during `Link step`.

On a `next.js` project this results in the following timings:
||before|after|
|--|--|--|
|yarn add leftpad| 16.47s | 13.44s|

**How did you fix it?**
<!-- A detailed description of your implementation. -->

I have added a `skipIntegrityCheck` option to `Fetcher` that can be used during Link step to not run integrity checks on the same files the second time. 

**Checklist**
<!--- Don't worry if you miss something, chores are automatically tested. -->
<!--- This checklist exists to help you remember doing the chores when you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).

<!-- See https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check -i` -->
- [x] I have set the packages that need to be released for my changes to be effective.

<!-- The "Testing chores" workflow validates that your PR follows our guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR might be missing. -->
- [x] I have verified that all automated PR checks pass.