@arcanis 2020-06-19
That's a significant improvement, but I'm not sure this is the right approach - the cache shouldn't have a singleton checksum cache (multiple caches may live in the same process). I'd prefer to disable the cache validation during linking.