@larixer 2020-06-19
@arcanis Thanks, I will look into how to do this the way you described